+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #5 Merging pages/components from different modules
 * #6 Upgrade to BQ 0.21
+* #7 Ignored paths
 
 ## 0.3